Extract PatchRequestToHandler into its own files
