+++ api/files.go
@@ -1,14 +1,10 @@
 package api
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"errors"
-	"io/ioutil"
 	"net/http"
-
-	"github.com/ONSdigital/log.go/v2/log"
 
 	"github.com/ONSdigital/dp-files-api/files"
 	"github.com/go-playground/validator"
@@ -51,13 +47,6 @@ type FilesCollection struct {
 	Items      []files.StoredRegisteredMetaData `json:"items"`
 }
 
-type PatchRequestHandlers struct {
-	UploadComplete   http.HandlerFunc
-	Published        http.HandlerFunc
-	Decrypted        http.HandlerFunc
-	CollectionUpdate http.HandlerFunc
-}
-
 type RegisterFileUpload func(ctx context.Context, metaData files.StoredRegisteredMetaData) error
 type MarkUploadComplete func(ctx context.Context, metaData files.FileEtagChange) error
 type GetFileMetadata func(ctx context.Context, path string) (files.StoredRegisteredMetaData, error)
@@ -66,53 +55,6 @@ type MarkFilePublished func(ctx context.Context, path string) error
 type MarkDecryptionComplete func(ctx context.Context, change files.FileEtagChange) error
 type UpdateCollectionID func(ctx context.Context, path, collectionID string) error
 type GetFilesMetadata func(ctx context.Context, collectionID string) ([]files.StoredRegisteredMetaData, error)
-
-func PatchRequestToHandler(handlers PatchRequestHandlers) http.HandlerFunc {
-	return func(w http.ResponseWriter, req *http.Request) {
-		stateMetaData, err := getStateMetadataFromRequest(req)
-
-		if err != nil {
-			writeError(w, buildErrors(err, "BadJsonEncoding"), http.StatusBadRequest)
-			return
-		}
-
-		if stateMetaData.CollectionID != nil && stateMetaData.State == nil {
-			handlers.CollectionUpdate.ServeHTTP(w, req)
-			return
-		}
-
-		switch *stateMetaData.State {
-		case files.StateUploaded:
-			handlers.UploadComplete.ServeHTTP(w, req)
-		case files.StatePublished:
-			handlers.Published.ServeHTTP(w, req)
-		case files.StateDecrypted:
-			handlers.Decrypted.ServeHTTP(w, req)
-		default:
-			log.Error(req.Context(), "InvalidStateChange", errors.New("invalid state change"), log.Data{"state": *stateMetaData.State})
-			writeError(w, buildErrors(errors.New("invalid state change"), "InvalidStateChange"), http.StatusBadRequest)
-		}
-	}
-}
-
-func setRequestBody(req *http.Request, requestBody []byte) {
-	req.Body = ioutil.NopCloser(bytes.NewBuffer(requestBody))
-}
-
-func getStateMetadataFromRequest(req *http.Request) (StateMetadata, error) {
-	stateMetaData := StateMetadata{}
-	requestBody, err := ioutil.ReadAll(req.Body)
-
-	setRequestBody(req, requestBody)
-
-	if err == nil {
-		requestBodyBuffer := bytes.NewBuffer(requestBody)
-		state := ioutil.NopCloser(requestBodyBuffer)
-		err = json.NewDecoder(state).Decode(&stateMetaData)
-	}
-
-	return stateMetaData, err
-}
 
 func HandlerUpdateCollectionID(updateCollectionID UpdateCollectionID) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
@@ -208,18 +150,6 @@ func HandlerGetFilesMetadata(getFilesMetadata GetFilesMetadata) http.HandlerFunc
 	}
 }
 
-func filesCollectionFromMetadata(f []files.StoredRegisteredMetaData) FilesCollection {
-	count := int64(len(f))
-	fc := FilesCollection{
-		Count:      count,
-		Limit:      count,
-		Offset:     0,
-		TotalCount: count,
-		Items:      f,
-	}
-	return fc
-}
-
 func HandlerRegisterUploadStarted(register RegisterFileUpload) http.HandlerFunc {
 	return func(w http.ResponseWriter, req *http.Request) {
 		rm, err := getRegisterMetadataFromRequest(req)
@@ -264,6 +194,18 @@ func HandleMarkUploadComplete(markUploaded MarkUploadComplete) http.HandlerFunc 
 
 		w.WriteHeader(http.StatusOK)
 	}
+}
+
+func filesCollectionFromMetadata(f []files.StoredRegisteredMetaData) FilesCollection {
+	count := int64(len(f))
+	fc := FilesCollection{
+		Count:      count,
+		Limit:      count,
+		Offset:     0,
+		TotalCount: count,
+		Items:      f,
+	}
+	return fc
 }
 
 func respondWithFilesCollectionJSON(w http.ResponseWriter, fc FilesCollection) error {
